rename data folder to store
