+++ src/components/PageFooter.tsx
@@ -2,9 +2,9 @@ import { Layout } from 'antd';
 import { useMemo } from 'react';
 import { useLocation } from 'react-router-dom';
 import { DEFAULT_TAB, TABS, TabName } from 'App';
-import useStore from 'data/useStore';
 import { Frameworks } from 'pages/frameworks/FrameworkData';
 import { overclocks } from 'pages/overclocks/OverclockData';
+import useStore from 'store/useStore';
 import { MinerColor } from 'utils/miner';
 import { MinerWeapons } from 'utils/weapons';
 import './PageFooter.css';
