+++ src/components/settingsModal/ResetProgress.tsx
@@ -2,7 +2,7 @@ import { Button, Col, Row, Space, Typography } from 'antd';
 import Avatar from 'antd/lib/avatar/avatar';
 import { SatchelCharge } from 'assets/other';
 import { MissionControlPortrait } from 'assets/portraits';
-import useStore from 'data/useStore';
+import useStore from 'store/useStore';
 const { Title, Text } = Typography;
 
 export default function ResetProgress(props: { hide: () => void }) {
