+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import './index.css';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Store from 'data/Store';
+import Store from 'store/Store';
 import App from './App';
 
 render(
