+++ src/pages/frameworks/MinerFrameworks.tsx
@@ -2,7 +2,7 @@
 import { Row } from 'antd';
 import React from 'react';
 import WeaponDivider from 'components/WeaponDivider';
-import useStore from 'data/useStore';
+import useStore from 'store/useStore';
 import { Miner } from 'utils/miner';
 import { MinerWeapon, MinerWeapons } from 'utils/weapons';
 import FrameworkCard from './FrameworkCard';
